backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-multipart
python-dotenv
